+++ packages/expo-notifications/ios/EXNotifications/Permissions/EXNotificationPermissionsModule.m
@@ -5,12 +5,14 @@
 #import <UMPermissionsInterface/UMPermissionsInterface.h>
 #import <UMPermissionsInterface/UMPermissionsMethodsDelegate.h>
 
+#import <EXNotifications/EXRemoteNotificationPermissionRequester.h>
 #import <EXNotifications/EXUserFacingNotificationsPermissionsRequester.h>
 
 @interface EXNotificationPermissionsModule ()
 
 @property (nonatomic, weak) id<UMPermissionsInterface> permissionsManager;
 @property (nonatomic, strong) EXUserFacingNotificationsPermissionsRequester *requester;
+@property (nonatomic, strong) EXRemoteNotificationPermissionRequester *legacyRemoteNotificationsRequester;
 
 @end
 
@@ -50,7 +52,11 @@ UM_EXPORT_METHOD_AS(requestPermissionsAsync,
 
 - (void)setModuleRegistry:(UMModuleRegistry *)moduleRegistry {
   _permissionsManager = [moduleRegistry getModuleImplementingProtocol:@protocol(UMPermissionsInterface)];
-  [UMPermissionsMethodsDelegate registerRequesters:@[_requester]
+  if (!_legacyRemoteNotificationsRequester) {
+    // TODO: Remove once we deprecate and remove "notifications" permission type
+    _legacyRemoteNotificationsRequester = [[EXRemoteNotificationPermissionRequester alloc] initWithUserNotificationPermissionRequester:_requester permissionPublisher:[moduleRegistry getSingletonModuleForName:@"RemoteNotificationPermissionPublisher"] withMethodQueue:self.methodQueue];
+  }
+  [UMPermissionsMethodsDelegate registerRequesters:@[_requester, _legacyRemoteNotificationsRequester]
                             withPermissionsManager:_permissionsManager];
 }